+++ src/Application/Bus/QueryDispatcher.php
@@ -109,7 +109,7 @@ class QueryDispatcher implements IQueryDispatcher
 
         foreach ($reflection->getAttributes(Through::class) as $attribute) {
             $instance = $attribute->newInstance();
-            $this->pipes[] = $instance->pipe;
+            $this->pipes = $instance->pipes;
         }
     }
 }